fixed staff card usages in projects/id and acocunts page; also restructured accountspage into its own page folder instead of keeping it in the components folder
